+++ examples/complete/main.tf
@@ -1,26 +1,33 @@
-### Supporting resources for complete example also showing module usage with external KMS Key
+### **NOTE**: This example should be run on management account
 module "kms_key" {
-  source                  = "boldlink/kms/aws"
-  version                 = "1.1.0"
-  description             = "Key used to encrypt/decrypt cloudTrail log files stored in S3."
-  create_kms_alias        = true
-  alias_name              = "alias/${local.name}-key-alias"
-  deletion_window_in_days = var.key_deletion_window_in_days
-  kms_policy              = local.kms_policy
-  tags                    = local.tags
+  source           = "boldlink/kms/aws"
+  version          = "1.1.0"
+  description      = "kms key for ${var.name}"
+  create_kms_alias = true
+  kms_policy       = local.org_kms_policy
+  alias_name       = "alias/${var.name}-key-alias"
+  tags             = local.tags
 }
 
-module "complete" {
+module "trail_bucket" {
+  source                 = "boldlink/s3/aws"
+  version                = "2.3.0"
+  bucket                 = var.name
+  bucket_policy          = data.aws_iam_policy_document.org_s3.json
+  sse_kms_master_key_arn = module.kms_key.arn
+  force_destroy          = true
+  tags                   = local.tags
+}
+
+module "aws_cloudtrail" {
   source                     = "../../"
-  name                       = local.name
+  name                       = var.name
   enable_log_file_validation = true
-  use_external_kms_key_id    = true
-  external_kms_key_id        = local.external_kms_key_id
-  bucket_name                = local.name
-  trail_name                 = local.name
   enable_logging             = true
+  kms_key_id                 = module.kms_key.arn
+  s3_bucket_name             = module.trail_bucket.bucket
+  is_organization_trail      = true
   tags                       = local.tags
-  depends_on                 = [module.kms_key]
 
   event_selectors = [
     {