+++ examples/complete/variables.tf
@@ -1,5 +1,19 @@
-variable "key_deletion_window_in_days" {
-  type        = number
-  description = "The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the KMS key. If you specify a value, it must be between 7 and 30, inclusive."
-  default     = 7
+variable "name" {
+  type        = string
+  description = "The name of the stack"
+  default     = "complete-trail-example"
+}
+
+variable "tags" {
+  type        = map(string)
+  description = "The tags to apply to the resources"
+  default = {
+    Environment        = "example"
+    "user::CostCenter" = "terraform-registry"
+    Department         = "DevOps"
+    Project            = "Examples"
+    Owner              = "Boldlink"
+    LayerName          = "cExample"
+    LayerId            = "cExample"
+  }
 }