+++ examples/organization_trail_with_external_bucket/outputs.tf
@@ -1,6 +1,0 @@
-output "outputs" {
-  value = [
-    module.aws_cloudtrail,
-  ]
-  description = "Output values for trail with external bucket"
-}