+++ main.tf
@@ -1,7 +1,6 @@
-### Cloud Trail
 resource "aws_cloudtrail" "main" {
-  name                          = local.trail_name
-  s3_bucket_name                = var.use_external_bucket ? var.s3_bucket_name : module.cloudtrail[0].id
+  name                          = var.name
+  s3_bucket_name                = var.s3_bucket_name
   s3_key_prefix                 = var.s3_key_prefix
   cloud_watch_logs_group_arn    = "${aws_cloudwatch_log_group.cloudtrail.arn}:*" # CloudTrail requires the Log Stream wildcard
   include_global_service_events = var.include_global_service_events
@@ -10,7 +9,7 @@ resource "aws_cloudtrail" "main" {
   enable_logging                = var.enable_logging
   is_multi_region_trail         = var.is_multi_region_trail
   is_organization_trail         = var.is_organization_trail
-  kms_key_id                    = var.use_external_kms_key_id ? var.external_kms_key_id : aws_kms_key.cloudtrail[0].arn
+  kms_key_id                    = var.kms_key_id
   sns_topic_name                = var.sns_topic_name
 
   ##### Use ONLY ONE of either event_selector or advanced_event_selector
@@ -69,58 +68,29 @@ resource "aws_cloudtrail" "main" {
   tags = var.tags
 }
 
-## Trails Bucket
-module "cloudtrail" {
-  count                  = var.use_external_bucket ? 0 : 1
-  source                 = "boldlink/s3/aws"
-  version                = "2.3.0"
-  bucket                 = local.bucket_name
-  force_destroy          = var.s3_bucket_force_destroy
-  versioning_status      = var.trail_bucket_versioning_enabled
-  bucket_policy          = var.is_organization_trail && var.use_external_bucket == false ? data.aws_iam_policy_document.org_s3.json : data.aws_iam_policy_document.s3.json
-  sse_kms_master_key_arn = var.use_external_kms_key_id ? var.external_kms_key_id : aws_kms_key.cloudtrail[0].arn
-  tags                   = var.tags
-}
-
 ### Log Group
 resource "aws_cloudwatch_log_group" "cloudtrail" {
-  name              = "/aws/cloudtrail/${local.name}"
+  name              = "/aws/cloudtrail/${var.name}"
   retention_in_days = var.log_retention_days
-  kms_key_id        = var.use_external_kms_key_id ? var.external_kms_key_id : aws_kms_key.cloudtrail[0].arn
+  kms_key_id        = var.kms_key_id
   tags              = var.tags
 }
 
 ## Role
 resource "aws_iam_role" "cloudtrail_cloudwatch" {
-  name               = local.name
+  name               = var.name
   assume_role_policy = data.aws_iam_policy_document.cloudtrail_assume_role.json
   tags               = var.tags
 }
 
 resource "aws_iam_policy" "cloudtrail_cloudwatch" {
-  name   = "${local.name}-policy"
+  name   = "${var.name}-policy"
   policy = data.aws_iam_policy_document.cloudtrail_cloudwatch_logs.json
   tags   = var.tags
 }
 
 resource "aws_iam_policy_attachment" "cloudwatch" {
-  name       = "${local.name}-policy-attachment"
+  name       = "${var.name}-policy-attachment"
   policy_arn = aws_iam_policy.cloudtrail_cloudwatch.arn
   roles      = [aws_iam_role.cloudtrail_cloudwatch.name]
-}
-
-## Trail KMS
-resource "aws_kms_key" "cloudtrail" {
-  count                   = var.use_external_kms_key_id ? 0 : 1
-  description             = "Key used to encrypt CloudTrail log files stored in S3."
-  deletion_window_in_days = var.key_deletion_window_in_days
-  enable_key_rotation     = true
-  policy                  = var.is_organization_trail ? data.aws_iam_policy_document.org_kms_policy.json : data.aws_iam_policy_document.kms_policy.json
-  tags                    = var.tags
-}
-
-resource "aws_kms_alias" "cloudtrail" {
-  count         = var.use_external_kms_key_id ? 0 : 1
-  name          = "alias/cloudtrail/${local.name}"
-  target_key_id = aws_kms_key.cloudtrail[0].arn
 }